sweet_shop_backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-asyncio
httpx
python multipart